scripts/mohan/requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
six==1.15.0
numpy==1.18.5
pandas
scikit-learn==0.22
joblib
memoized-property
mlflow
s3fs
gcsfs
google-cloud-storage
termcolor
#git+https://github.com/Maluuba/nlg-eval.git@master
torchvision
TensorFlow
Keras
pillow
tqdm
time